# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SWAPI Laravel API",
        "description": "API REST que integra la Star Wars API (SWAPI) con caché, reintentos y DTOs. Gestiona planetas y habitantes localmente con relaciones Eloquent.",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API v1 – local"
        }
    ],
    "paths": {
...
        "/naves": {
            "get": {
                "tags": [
                    "Naves Locales"
                ],
                "summary": "Listar naves locales",
                "operationId": "285079c05cc9a5f527c565add57f6e29",
                "responses": {
                    "200": {
                        "description": "Lista paginada de naves"
                    },
                    "401": {
                        "description": "API Key inválida"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Naves Locales"
                ],
                "summary": "Crear una nave local",
                "operationId": "32521b21119326f7628cd01173833837",
                "responses": {
                    "201": {
                        "description": "Nave creada"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "401": {
                        "description": "API Key inválida"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/naves/{id}": {
            "get": {
                "tags": [
                    "Naves Locales"
                ],
                "summary": "Obtener una nave local con sus pilotos y películas",
                "operationId": "e43db787382420fe887e462ea8ef8037",
                "responses": {
                    "200": {
                        "description": "Nave encontrada"
                    },
                    "404": {
                        "description": "Nave no encontrada"
                    },
                    "401": {
                        "description": "API Key inválida"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Naves Locales"
                ],
                "summary": "Actualizar una nave local",
                "operationId": "6d5fc361ea552797d5aa3b8523c6b4c6",
                "responses": {
                    "200": {
                        "description": "Nave actualizada"
                    },
                    "404": {
                        "description": "Nave no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "401": {
                        "description": "API Key inválida"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Naves Locales"
                ],
                "summary": "Eliminar una nave local",
                "operationId": "ad22820f5378053c3e191c55e86bf5a2",
                "responses": {
                    "200": {
                        "description": "Nave eliminada"
                    },
                    "404": {
                        "description": "Nave no encontrada"
                    },
                    "401": {
                        "description": "API Key inválida"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "API Key requerida. Configura API_KEY en tu .env y envíala como header X-API-KEY.",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
...
}